Divide files by role
